Split daemon server helpers
